+++ src/platform_impl/macos/window.rs
@@ -1,3 +1,5 @@
+#![allow(clippy::unnecessary_cast)]
+
 use std::{
     collections::VecDeque,
     f64, ops,
@@ -225,7 +227,7 @@ impl WinitWindow {
                 Some(screen) => screen.frame(),
                 None => {
                     let scale_factor = NSScreen::main()
-                        .map(|screen| screen.backingScaleFactor())
+                        .map(|screen| screen.backingScaleFactor() as f64)
                         .unwrap_or(1.0);
                     let (width, height) = match attrs.inner_size {
                         Some(size) => {
@@ -490,7 +492,7 @@ impl WinitWindow {
     pub fn outer_position(&self) -> Result<PhysicalPosition<i32>, NotSupportedError> {
         let frame_rect = self.frame();
         let position = LogicalPosition::new(
-            frame_rect.origin.x,
+            frame_rect.origin.x as f64,
             util::bottom_left_to_top_left(frame_rect),
         );
         let scale_factor = self.scale_factor();
@@ -500,7 +502,7 @@ impl WinitWindow {
     pub fn inner_position(&self) -> Result<PhysicalPosition<i32>, NotSupportedError> {
         let content_rect = self.contentRectForFrameRect(self.frame());
         let position = LogicalPosition::new(
-            content_rect.origin.x,
+            content_rect.origin.x as f64,
             util::bottom_left_to_top_left(content_rect),
         );
         let scale_factor = self.scale_factor();
@@ -516,7 +518,7 @@ impl WinitWindow {
     #[inline]
     pub fn inner_size(&self) -> PhysicalSize<u32> {
         let frame = self.contentView().frame();
-        let logical: LogicalSize<f64> = (frame.size.width, frame.size.height).into();
+        let logical: LogicalSize<f64> = (frame.size.width as f64, frame.size.height as f64).into();
         let scale_factor = self.scale_factor();
         logical.to_physical(scale_factor)
     }
@@ -524,7 +526,7 @@ impl WinitWindow {
     #[inline]
     pub fn outer_size(&self) -> PhysicalSize<u32> {
         let frame = self.frame();
-        let logical: LogicalSize<f64> = (frame.size.width, frame.size.height).into();
+        let logical: LogicalSize<f64> = (frame.size.width as f64, frame.size.height as f64).into();
         let scale_factor = self.scale_factor();
         logical.to_physical(scale_factor)
     }
@@ -721,7 +723,7 @@ impl WinitWindow {
 
     #[inline]
     pub fn scale_factor(&self) -> f64 {
-        self.backingScaleFactor()
+        self.backingScaleFactor() as f64
     }
 
     #[inline]
